same update\fix + update swagger ver 2.1.1

+++ controllers/columns/deleteColumn.js
@@ -6,14 +6,26 @@ import Column from "../../models/column.js";
 
 const deleteColumn = ctrlWrapper(async (req, res) => {
   const { id } = req.params;
+  const { user } = req;
 
-  const resultDelete = await Column.findByIdAndDelete(id);
-  if (!resultDelete) {
+  const column = await Column.findById(id);
+  if (!column) {
     throw HttpError(404, `Column with id=${id} not found`);
   }
 
+  const board = await Board.findById(column.board);
+  if (!board) {
+    throw HttpError(404, `Board for column with id=${id} not found`);
+  }
+
+  const isOwner = board.owners.some((ownerId) => ownerId.equals(user._id));
+  if (!isOwner) {
+    throw HttpError(403, `You are not authorized to delete this column`);
+  }
+
+  const columnDelete = await Column.findByIdAndDelete(id);
   await Board.findByIdAndUpdate(
-    resultDelete.board,
+    columnDelete.board,
     { $pull: { columns: id } },
     { new: true }
   );
